Add files via upload
This allows the user to predict stock prices on the basis of rainfall in the Gulf of Mexico Area for a given number of years. The rainfall data is taken from here: http://rainsphere.eng.uci.edu 
In a few test runs (for example CL=F, BZ=F, ADM), no correlation was found. However, this model will become the foundation for future models that do make conclusive or close-to-accurate predictions on the basis of various weather parameters. Another limitation is that, right now, storms, hurricanes, etc. are not considered. Considering this information would probably make the predictions (at least for stocks like CL=F) more accurate. Future scope includes replicating this for rainfall in India and cotton prices, or rainfall and sugar cane prices.